Crate fo_sim

A Forward Observer (FO) Simulator

This crate provides an async FO Simulator built on tokio. The crate will listen for and talk to any FDC over the fo_fdc_comms message interface.

Modules

fo_fdc_commhandler

Contains the message types that the FO sim can send/receive, as well as the communication tasks (send and receive)

state_machine

Provides the functions and enums for maintaining the FO state machine.

Functions

main

Entry function