Module fo_sim::state_machine

Provides the functions and enums for maintaining the FO state machine.

The FO State Machine uses the top level FoState for representing the state of the FO. The workhorse of this module, and the intended component for use is the state_machine_loop, which provides an async function for use in a runtime.

Enums

ConnectedState

Representation of the Connectedstate of an FO.

FoState

Representation of the top-level state of a Forward Observer

Functions

state_machine_loop

Asynchronous executor loop for managing the state machine.